% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMat.R
\name{classMat}
\alias{classMat}
\title{Classification matrix estimating accuracy of model.}
\usage{
classMat(comb)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}
}
\value{
Matrix with \% correct predictions from training data.
}
\description{
Classification matrix estimating accuracy of model.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
COUNT <- "copy"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2
COPY_RANGE <- "1 99"

listPos <- tsvDir(system.file("extdata", "Post", package="iCAT"))
listNeg <- tsvDir(system.file("extdata", "Pre", package="iCAT"))

naive <- readTrn(listNeg, FIELD, COUNT, COPY_RANGE, "naive")
vaccs <- readTrn(listPos, FIELD, COUNT, COPY_RANGE, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, COUNT, COPY_RANGE, P_CUTOFF, MIN_PUBLIC, NULL)
classMat(mod)
}
