% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLib.R
\name{getLib}
\alias{getLib}
\title{Dataframe of TARSs library.}
\usage{
getLib(comb)
}
\arguments{
\item{comb}{composed list returned from train() function}
}
\value{
Dataframe containing the significant target-associated receptor sequences.
}
\description{
Dataframe of TARSs library.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
COUNT <- "count (templates)"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2
COPY_RANGE <- "1 99"

listPos <- tsvDir(system.file("extdata", "Post", package="iCAT"))
listNeg <- tsvDir(system.file("extdata", "Pre", package="iCAT"))

naive <- readTrn(listNeg, FIELD, COUNT, COPY_RANGE, "naive")
vaccs <- readTrn(listPos, FIELD, COUNT, COPY_RANGE, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, COUNT, COPY_RANGE, P_CUTOFF, MIN_PUBLIC, NULL)
getLib(mod)
}
