% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCATinteractive.R
\name{iCATinteractive}
\alias{iCATinteractive}
\title{Launch interactive interface}
\usage{
iCATinteractive()
}
\description{
Launch interactive interface
}
\examples{
if(interactive()) {iCATinteractive()}
}
