% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Predict the exposure status of an independent sample.}
\usage{
pred(comb, indpt, names, field, count, copyrange)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}

\item{indpt}{Vector of independent samples file paths.}

\item{names}{Vector of labels for independent samples.}

\item{field}{String containing the column or columns (space-delimited) of interest.}

\item{count}{String containing the column name for colontype counts.}

\item{copyrange}{Integer Vector of the min and max copy of a sequence, within a sample, to be considered.}
}
\value{
Matrix with \% correct predictions from training data.
}
\description{
Predict the exposure status of an independent sample.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
COUNT <- "copy"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2
COPY_RANGE <- "1 99"


listPos <- tsvDir(system.file("extdata", "Post", package="iCAT"))
listNeg <- tsvDir(system.file("extdata", "Pre", package="iCAT"))

naive <- readTrn(listNeg, FIELD, COUNT, COPY_RANGE, "naive")
vaccs <- readTrn(listPos, FIELD, COUNT, COPY_RANGE, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, COUNT, COPY_RANGE, P_CUTOFF, MIN_PUBLIC, NULL)
pred(mod, system.file("extdata", "Post", "post0.tsv", package="iCAT"), "unknown-sample-label", FIELD, COUNT, COPY_RANGE)
}
