% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{processFiles}
\alias{processFiles}
\title{Internal function for transforming input files.}
\usage{
processFiles(x, field)
}
\arguments{
\item{x}{Input file.}

\item{field}{String containing the column or columns (space-delimited) of interest.}
}
\description{
Internal function for transforming input files.
}
