% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{savePlot}
\alias{savePlot}
\title{Save plot to pdf.}
\usage{
savePlot(file, plotIn)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}
}
\description{
Save plot to pdf.
}
