% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Main training function.}
\usage{
train(negatives, positives, prelist, postlist, field, count, copyrange,
  pcut, minpublic, updateProgress = NULL)
}
\arguments{
\item{negatives}{Dataframe of sequence frequencies in negative samples.}

\item{positives}{Datafram of sequence frequencies in positive samples.}

\item{prelist}{Vector of negative training samples.}

\item{postlist}{Vector of positive training samples.}

\item{field}{String containing the column or columns (space-delimited) of interest.}

\item{count}{String containing the column name for colontype counts.}

\item{copyrange}{Integer Vector of the min and max copy of a sequence, within a sample, to be considered.}

\item{pcut}{P-value threshold for fisher-exact test.}

\item{minpublic}{Sequence frequency threshold to be considered.}

\item{updateProgress}{Function for updating a progress bar in a Shiny interface.}
}
\value{
List containing both negtive (n) and positive (v) clonotype percentages.
}
\description{
Main training function.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
COUNT <- "count (templates)"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2
COPY_RANGE <- "1 99"

listPos <- tsvDir(system.file("extdata", "Post", package="iCAT"))
listNeg <- tsvDir(system.file("extdata", "Pre", package="iCAT"))

naive <- readTrn(listNeg, FIELD, COUNT, COPY_RANGE, "naive")
vaccs <- readTrn(listPos, FIELD, COUNT, COPY_RANGE, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, COUNT, COPY_RANGE, P_CUTOFF, MIN_PUBLIC, NULL)
}
