% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvDir.R
\name{tsvDir}
\alias{tsvDir}
\title{List TSV files in a directory.}
\usage{
tsvDir(dir)
}
\arguments{
\item{dir}{String of directory.}
}
\value{
List of *.tsv files in the directory.
}
\description{
List TSV files in a directory.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
listPos <- tsvDir(system.file("extdata", "Post", package="iCAT"))
listNeg <- tsvDir(system.file("extdata", "Pre", package="iCAT"))

naive <- readTrn(listNeg, FIELD, "naive")
vaccs <- readTrn(listPos, FIELD, "vacc")  
}
