% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMat.R
\name{classMat}
\alias{classMat}
\title{Classification matrix estimating accuracy of model.}
\usage{
classMat(comb)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}
}
\value{
Matrix with \% correct predictions from training data.
}
\description{
Classification matrix estimating accuracy of model.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2

listPos <- tsvDir("iCAT/extdata/Post/")
listNeg <- tsvDir("iCAT/extdata/Pre/")

naive <- readTrn(listNeg, FIELD, "naive")
vaccs <- readTrn(listPos, FIELD, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, P_CUTOFF, MIN_PUBLIC, NULL)
classMat(mod)
}
