% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMat.R
\name{classMat}
\alias{classMat}
\title{Classification matrix estimating accuracy of model.}
\usage{
classMat(comb)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}
}
\value{
Matrix with % correct predictions from training data.
}
\description{
Classification matrix estimating accuracy of model.
}
\examples{
comb <- train(...)
classMat(comb)
}
