% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLib.R
\name{getLib}
\alias{getLib}
\title{Dataframe of TARSs library.}
\usage{
getLib(comb)
}
\arguments{
\item{comb}{composed list returned from train() function}
}
\value{
Dataframe containing the significant target-associated receptor sequences.
}
\description{
Dataframe of TARSs library.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2

listPos <- tsvDir("path/to/positve/samples/")
listNeg <- tsvDir("path/to/negative/samples/")

naive <- readTrn(listNeg, FIELD, "naive")
vaccs <- readTrn(listPos, FIELD, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, P_CUTOFF, MIN_PUBLIC, NULL)
getLib(mod)
}
