% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLib.R
\name{getLib}
\alias{getLib}
\title{Dataframe of TARSs library.}
\usage{
getLib(comb)
}
\arguments{
\item{comb}{composed list returned from train() function}
}
\value{
Dataframe containing the significant target-associated receptor sequences.
}
\description{
Dataframe of TARSs library.
}
