% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{Plot histogram of clonotype frequencies in negative and positive samples.}
\usage{
plotHist(comb)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}
}
\description{
Plot histogram of clonotype frequencies in negative and positive samples.
}
\examples{
comb <- train(...)
plotHist(comb)
}
