% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Predict the exposure status of an independent sample.}
\usage{
pred(comb, indpt, names, field)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}

\item{indpt}{Vector of independent samples file paths.}

\item{names}{Vector of labels for independent samples.}

\item{field}{String containing the column or columns (space-delimited) of interest.}
}
\value{
Matrix with \% correct predictions from training data.
}
\description{
Predict the exposure status of an independent sample.
}
\examples{
FIELD <- "vGeneName aminoAcid jGeneName"
P_CUTOFF <- 0.1
MIN_PUBLIC <- 2

listPos <- tsvDir("iCAT/extdata/Post/")
listNeg <- tsvDir("iCAT/extdata/Pre/")

naive <- readTrn(listNeg, FIELD, "naive")
vaccs <- readTrn(listPos, FIELD, "vacc")  

mod <- train(naive, vaccs, listNeg, listPos, FIELD, P_CUTOFF, MIN_PUBLIC, NULL)
pred(mod, "iCAT/extdata/Post/post0.tsv", "unknown-sample-label", FIELD)
}
