% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Predict the exposure status of an independent sample.}
\usage{
pred(comb, iccs, indpt, names, field)
}
\arguments{
\item{comb}{List containing both negtive (n) and positive (v) clonotype percentages.}

\item{iccs}{Dataframe of sequence frequencies in independent sample.}

\item{indpt}{Vector of independent samples file paths.}

\item{names}{Vector of labels for independent samples.}

\item{field}{String containing the column or columns (space-delimited) of interest.}
}
\value{
Matrix with % correct predictions from training data.
}
\description{
Predict the exposure status of an independent sample.
}
\examples{
comb <- train(...)
classMat(comb)
}
