% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrn.R
\name{readTrn}
\alias{readTrn}
\title{Read input training data.}
\usage{
readTrn(list, field, posOrNeg)
}
\arguments{
\item{list}{Vector of training samples.}

\item{field}{String containing the column or columns (space-delimited) of interest.}

\item{posOrNeg}{String indicating whether the data is negative or positive.}
}
\value{
Dataframe containing unique sequences and their frequencies in the samples.
}
\description{
Read input training data.
}
