% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Main training function.}
\usage{
train(negatives, positives, prelist, postlist, field, pcut, minpublic,
  updateProgress)
}
\arguments{
\item{negatives}{Dataframe of sequence frequencies in negative samples.}

\item{positives}{Datafram of sequence frequencies in positive samples.}

\item{prelist}{Vector of negative training samples.}

\item{postlist}{Vector of positive training samples.}

\item{field}{String containing the column or columns (space-delimited) of interest.}

\item{pcut}{P-value threshold for fisher-exact test.}

\item{minpublic}{Sequence frequency threshold to be considered.}

\item{updateProgress}{Function for updating a progress bar in a Shiny interface.}
}
\value{
List containing both negtive (n) and positive (v) clonotype percentages.
}
\description{
Main training function.
}
